refactored anuTests
