+++ src/test/java/api/anuTests/GetAllBrandsList.java
@@ -1,7 +1,5 @@
-package api.endpoints.anu;
+package api.anuTests;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.builder.ResponseSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
