--- src/test/java/api/endpoints/anu/Brand.java
@@ -1,4 +1,0 @@
-package api.endpoints.anu;
-
-public class Brand {
-}
